requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
exceptiongroup==1.1.2
gunicorn==20.1.0
iniconfig==2.0.0
jose==1.0.0
packaging==23.1
pluggy==1.2.0
pyasn1==0.5.0
PyJWT==1.7.1
pytest==7.4.0
python-jose==3.3.0
pytz==2023.3
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
